Add mobile lookup
I know it lacks test but I am short on time. There are few things that
could be improved as well, for example extracting MessageBird client to
a separate instance, so Messenger and Msisdn could use it (and mock in
tests as well).
I don't know what time do we have to spend on improvements here, so any
advice would be more than appreciated.